requirements.txt
cython
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI&egg=pycocotools
submitit
torch>=1.5.0
torchvision>=0.6.0
git+https://github.com/cocodataset/panopticapi.git#egg=panopticapi
scipy~=1.15.2
onnx
onnxruntime~=1.21.0
matplotlib~=3.10.1
packaging~=24.2
pandas~=2.2.3
numpy~=1.26.4
opencv-python~=4.11.0.86
pillow~=11.1.0